+++ src/main/java/com/utp/viacosta/model/ClienteModel.java
@@ -18,7 +18,7 @@ public class ClienteModel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_cliente")
-    private Long idCliente;
+    private Integer idCliente;
     private String nombre;
     private String apellido;
     private String dni;
